chore(animation): optimize files+
