+++ src/packages/interactive/animation/src/ali-miniapp/index.ts
@@ -1,11 +1,9 @@
-import { Animation, AnimationOptions } from '../types';
-import { CONTAINER_NAME } from '@utils/constant';
-import { normalizeOptions } from '../common';
+import createAnimation from './createAnimation';
 
-function createAnimation(options?: AnimationOptions): Animation {
-  return my.createAnimation(
-    normalizeOptions(options, CONTAINER_NAME.ALIPAY),
-  );
-}
+export {
+  createAnimation,
+};
 
-export default createAnimation;
+export default {
+  createAnimation,
+};
