+++ src/packages/interactive/animation/src/bytedance-microapp/createAnimation.ts
@@ -1,0 +1,11 @@
+import { Animation, AnimationOptions } from '../types';
+import { CONTAINER_NAME } from '@utils/constant';
+import { normalizeOptions } from '../common';
+
+function createAnimation(options?: AnimationOptions): Animation {
+  return tt.createAnimation(
+    normalizeOptions(options, CONTAINER_NAME.BYTE),
+  );
+}
+
+export default createAnimation;
