+++ src/packages/interactive/animation/src/index.ts
@@ -5,18 +5,20 @@ import weChatModule from './wechat-miniprogram/index';
 import bytedanceModule from './bytedance-microapp/index';
 import { AnimationOptions } from './types';
 
-function createAnimation(options?: AnimationOptions) {
+export function createAnimation(options?: AnimationOptions) {
   if (isWeChatMiniProgram) {
-    return weChatModule(options);
+    return weChatModule.createAnimation(options);
   } else if (isByteDanceMicroApp) {
-    return bytedanceModule(options);
+    return bytedanceModule.createAnimation(options);
   } else if (isMiniApp) {
-    return aliMiniAppModule(options);
+    return aliMiniAppModule.createAnimation(options);
   } else if (isWeb) {
-    return webModule(options);
+    return webModule.createAnimation(options);
   } else {
     throw new Error('@uni/apis：createAnimation 暂不支持');
   }
 }
 
-export default createAnimation;
+export default {
+  createAnimation,
+};
