+++ src/packages/interactive/animation/src/web/createAnimation.ts
@@ -1,0 +1,8 @@
+import { Animation, AnimationOptions } from '../types';
+import AnimationImpl from './AnimationImpl';
+
+function createAnimation(options?: AnimationOptions): Animation {
+  return new AnimationImpl(options);
+}
+
+export default createAnimation;
